Module aoc_utils::maze

Implementations of 2D mazes.

Structs

HashMapMaze

A 2D maze where each location and value in the maze is stored in a HashMap. The idea here is that each key in the hashmap uniquely identifies a location in a 2D grid. Therefore, a natural choice for the generic type L is (usize, usize)

HashMapOccupiedMaze

Wraps a HashMapMaze with a current location L to enable a set of associated functions for mutably moving around in the maze.

Enums

Block

Enum for defining locations in a maze which are accessible or inaccessible.

MazeError

Error type for 2D mazes and associated functions.

Traits

NavigableMaze

This trait describes a navigable maze, where each generic Self::Location type has optionally has a loc above, below, to the right, and to the left of it.

SingleOccupantMaze

This supertrait is a describes a NavigableMaze with state representing the current location of a single occupant of the maze.