Trait aoc_utils::maze::NavigableMaze

pub trait NavigableMaze {
    type Location;

    fn loc_above(&self, loc: Self::Location) -> Result<Self::Location, MazeError>;
    fn loc_below(&self, loc: Self::Location) -> Result<Self::Location, MazeError>;
    fn loc_right(&self, loc: Self::Location) -> Result<Self::Location, MazeError>;
    fn loc_left(&self, loc: Self::Location) -> Result<Self::Location, MazeError>;
}

This trait describes a navigable maze, where each generic Self::Location type has optionally has a loc above, below, to the right, and to the left of it.

Required Associated Types

type Location

Represents a location in a 2D maze.

Required Methods

fn loc_above(&self, loc: Self::Location) -> Result<Self::Location, MazeError>

Find the location in the 2D maze above loc

fn loc_below(&self, loc: Self::Location) -> Result<Self::Location, MazeError>

Find the location in the 2D maze below loc

fn loc_right(&self, loc: Self::Location) -> Result<Self::Location, MazeError>

Find the location in the 2D maze to the right of loc

fn loc_left(&self, loc: Self::Location) -> Result<Self::Location, MazeError>

Find the location in the 2D maze to the left of loc