Trait aoc_utils::maze::SingleOccupantMaze

pub trait SingleOccupantMaze: NavigableMaze {
    fn move_up(&mut self) -> Result<(), MazeError>;
    fn move_down(&mut self) -> Result<(), MazeError>;
    fn move_right(&mut self) -> Result<(), MazeError>;
    fn move_left(&mut self) -> Result<(), MazeError>;
}

This supertrait is a describes a NavigableMaze with state representing the current location of a single occupant of the maze.

Required Methods

fn move_up(&mut self) -> Result<(), MazeError>

Move the occupant of the maze up one

fn move_down(&mut self) -> Result<(), MazeError>

Move the occupant of the maze down one

fn move_right(&mut self) -> Result<(), MazeError>

Move the occupant of the maze to the right one

fn move_left(&mut self) -> Result<(), MazeError>

Move the occupant of the maze to the left one